[Opt](flexible partial update) let `SegmentWriter` do not support flexible partial update (#41950)
Considering that `SegmentWriter` will be removed in the near future, we
stop to support flexible partial update on `SegmentWriter` to reduce
maintenance costs
