+++ regression-test/data/unique_with_mow_p0/flexible/test_f_segment_writer.out
@@ -1,0 +1,17 @@
+-- This file is automatically generated. You should know what you did if you want to edit this
+-- !sql --
+0	0	0	0	0	0	
+1	1	1	1	1	1	
+2	2	2	2	2	2	
+3	3	3	3	3	3	
+4	4	4	4	4	4	
+5	5	5	5	5	5	
+
+-- !sql --
+0	0	0	0	0	0	
+1	1	1	1	1	1	
+2	2	2	2	2	2	
+3	3	3	3	3	3	
+4	4	4	4	4	4	
+5	5	5	5	5	5	
+
