--- regression-test/data/unique_with_mow_p0/flexible/test_f_seq_col_no_vertical_segment_writer.out
@@ -1,271 +1,0 @@
--- This file is automatically generated. You should know what you did if you want to edit this
--- !seq_map_no_default_val1 --
-0	0	0	0	0	0	0	
-1	1	1	1	1	10	10	
-2	2	2	2	2	20	20	
-3	3	3	3	3	30	30	
-4	4	4	4	4	40	40	
-5	5	5	5	5	50	50	
-
--- !seq_map_no_default_val2 --
-0	0	0	0	0	0	0	
-1	1	1	1	1	10	10	
-2	2	222	2	2	25	25	1,3,4,6
-3	3	3	333	3	30	30	1,2,4,5,6,9
-4	411	\N	433	444	50	50	6
-5	5	5	5	5	50	50	
-6	611	9876	633	1234	\N	\N	2,4,5,6,9
-7	\N	9876	733	1234	300	300	1,2,4,6
-
--- !seq_map_has_default_val1 --
-0	0	0	0	0	0	0	
-1	1	1	1	1	10	10	
-2	2	2	2	2	20	20	
-3	3	3	3	3	30	30	
-4	4	4	4	4	40	40	
-5	5	5	5	5	50	50	
-
--- !seq_map_has_default_val2 --
-0	0	0	0	0	0	0	
-1	1	1	1	1	10	10	
-2	2	222	2	2	25	25	1,3,4,6
-3	3	3	333	3	30	30	1,2,4,5,6,9
-4	411	\N	433	444	50	50	6
-5	5	5	5	5	50	50	
-6	611	9876	633	1234	31	31	2,4,5,6,9
-7	\N	9876	733	1234	300	300	1,2,4,6
-
--- !seq_type_col1 --
-0	0	0	0	0	0	0	
-1	1	1	1	1	1	10	
-2	2	2	2	2	2	20	
-3	3	3	3	3	3	30	
-4	4	4	4	4	4	40	
-5	5	5	5	5	5	50	
-
--- !seq_type_col2 --
-0	0	0	0	0	0	0	
-1	1	1	1	1	1	10	
-2	2	222	2	2	25	25	1,3,4,6
-3	3	3	333	3	3	30	1,2,4,5,6,9
-4	411	\N	433	444	50	50	6
-5	5	5	5	5	5	50	
-6	611	9876	633	1234	\N	\N	2,4,5,6,9
-7	\N	9876	733	1234	300	300	1,2,4,6
-
--- !seq_type_col_multi_rows_1 --
-0	0	0	0	0	0	0	
-1	1	1	1	1	1	10	
-2	2	2	2	2	2	20	
-3	3	3	3	3	3	30	
-4	4	4	4	4	4	40	
-5	5	5	5	5	5	50	
-
--- !seq_type_col_multi_rows_2 --
-0	0	0	0	0	0	0	
-1	888	888	1	1	1	20	3,4,5,6
-2	2	2	777	777	777	20	1,2,6,9
-3	3	3	3	3	3	30	
-4	4	4	4	4	4	40	
-5	5	5	5	5	5	50	
-6	666	666	0	1234	\N	30	3,4,5,6
-
--- !seq_type_col_multi_rows_3 --
-0	0	0	0	0	0	0	
-1	888	888	1	1	1	20	3,4,5,6
-2	2	2	777	777	777	20	1,2,6,9
-3	777	777	555	555	555	53	6
-4	123	123	123	987	123	40	6,9
-5	5	5	5	5	5	50	
-6	666	666	0	1234	\N	30	3,4,5,6
-7	666	666	0	1234	\N	30	3,4,5,6
-
--- !seq_map_col_no_default_val_multi_rows_1 --
-0	0	0	0	0	0	0	0	
-1	1	1	1	1	1	10	10	
-2	2	2	2	2	2	20	20	
-3	3	3	3	3	3	30	30	
-4	4	4	4	4	4	40	40	
-5	5	5	5	5	5	50	50	
-
--- !seq_map_col_no_default_val_multi_rows_2 --
-0	0	0	0	0	0	0	0	
-1	888	888	1	1	1	20	20	3,4,5,7
-2	2	2	777	777	777	20	20	1,2,6,7,10
-3	3	3	3	3	3	30	30	
-4	4	4	4	4	4	40	40	
-5	5	5	5	5	5	50	50	
-6	666	666	0	1234	\N	30	30	3,4,5,7
-
--- !seq_map_col_no_default_val_multi_rows_3 --
-0	0	0	0	0	0	0	
-1	888	888	1	1	1	20	3,4,5,7
-2	2	2	777	777	777	20	1,2,6,7,10
-3	777	777	555	555	555	53	7
-4	123	123	123	987	123	40	6,7,10
-5	5	5	5	5	5	50	
-6	666	666	0	1234	\N	30	3,4,5,7
-7	666	666	0	1234	\N	70	3,4,5,7
-
--- !seq_map_col_has_default_val_multi_rows_1 --
-0	0	0	0	0	0	0	0	
-1	1	1	1	1	1	10	10	
-2	2	2	2	2	2	20	20	
-3	3	3	3	3	3	30	30	
-4	4	4	4	4	4	40	40	
-5	5	5	5	5	5	50	50	
-
--- !seq_map_col_has_default_val_multi_rows_2 --
-0	0	0	0	0	0	0	0	
-1	888	888	1	1	1	20	20	3,4,5,7
-2	2	2	777	777	777	20	20	1,2,6,7,10
-3	3	3	3	3	3	30	30	
-4	4	4	4	4	4	40	40	
-5	5	5	5	5	5	50	50	
-6	123	123	0	666	666	60	60	3,6,7,10
-
--- !seq_map_col_has_default_val_multi_rows_3 --
-0	0	0	0	0	0	0	
-1	888	888	1	1	1	20	3,4,5,7
-2	2	2	777	777	777	20	1,2,6,7,10
-3	777	777	555	555	555	53	7
-4	123	123	123	987	123	40	6,7,10
-5	5	5	5	5	5	50	
-6	123	123	0	666	666	60	3,6,7,10
-7	666	666	0	1234	\N	70	3,4,5,7
-
--- !seq_map_no_default_val1 --
-0	0	0	0	0	0	0	
-1	1	1	1	1	10	10	
-2	2	2	2	2	20	20	
-3	3	3	3	3	30	30	
-4	4	4	4	4	40	40	
-5	5	5	5	5	50	50	
-
--- !seq_map_no_default_val2 --
-0	0	0	0	0	0	0	
-1	1	1	1	1	10	10	
-2	2	222	2	2	25	25	1,3,4,6
-3	3	3	333	3	30	30	1,2,4,5,6,10
-4	411	\N	433	444	50	50	6
-5	5	5	5	5	50	50	
-6	611	9876	633	1234	\N	\N	2,4,5,6,10
-7	\N	9876	733	1234	300	300	1,2,4,6
-
--- !seq_map_has_default_val1 --
-0	0	0	0	0	0	0	
-1	1	1	1	1	10	10	
-2	2	2	2	2	20	20	
-3	3	3	3	3	30	30	
-4	4	4	4	4	40	40	
-5	5	5	5	5	50	50	
-
--- !seq_map_has_default_val2 --
-0	0	0	0	0	0	0	
-1	1	1	1	1	10	10	
-2	2	222	2	2	25	25	1,3,4,6
-3	3	3	333	3	30	30	1,2,4,5,6,10
-4	411	\N	433	444	50	50	6
-5	5	5	5	5	50	50	
-6	611	9876	633	1234	31	31	2,4,5,6,10
-7	\N	9876	733	1234	300	300	1,2,4,6
-
--- !seq_type_col1 --
-0	0	0	0	0	0	0	
-1	1	1	1	1	1	10	
-2	2	2	2	2	2	20	
-3	3	3	3	3	3	30	
-4	4	4	4	4	4	40	
-5	5	5	5	5	5	50	
-
--- !seq_type_col2 --
-0	0	0	0	0	0	0	
-1	1	1	1	1	1	10	
-2	2	222	2	2	25	25	1,3,4,6
-3	3	3	333	3	3	30	1,2,4,5,6,10
-4	411	\N	433	444	50	50	6
-5	5	5	5	5	5	50	
-6	611	9876	633	1234	\N	\N	2,4,5,6,10
-7	\N	9876	733	1234	300	300	1,2,4,6
-
--- !seq_type_col_multi_rows_1 --
-0	0	0	0	0	0	0	
-1	1	1	1	1	1	10	
-2	2	2	2	2	2	20	
-3	3	3	3	3	3	30	
-4	4	4	4	4	4	40	
-5	5	5	5	5	5	50	
-
--- !seq_type_col_multi_rows_2 --
-0	0	0	0	0	0	0	
-1	888	888	1	1	1	20	3,4,5,6
-2	2	2	777	777	777	20	1,2,6,10
-3	3	3	3	3	3	30	
-4	4	4	4	4	4	40	
-5	5	5	5	5	5	50	
-6	666	666	0	1234	\N	30	3,4,5,6
-
--- !seq_type_col_multi_rows_3 --
-0	0	0	0	0	0	0	
-1	888	888	1	1	1	20	3,4,5,6
-2	2	2	777	777	777	20	1,2,6,10
-3	777	777	555	555	555	53	6
-4	123	123	123	987	123	40	6,10
-5	5	5	5	5	5	50	
-6	666	666	0	1234	\N	30	3,4,5,6
-7	666	666	0	1234	\N	30	3,4,5,6
-
--- !seq_map_col_no_default_val_multi_rows_1 --
-0	0	0	0	0	0	0	0	
-1	1	1	1	1	1	10	10	
-2	2	2	2	2	2	20	20	
-3	3	3	3	3	3	30	30	
-4	4	4	4	4	4	40	40	
-5	5	5	5	5	5	50	50	
-
--- !seq_map_col_no_default_val_multi_rows_2 --
-0	0	0	0	0	0	0	0	
-1	888	888	1	1	1	20	20	3,4,5,7
-2	2	2	777	777	777	20	20	1,2,6,7,11
-3	3	3	3	3	3	30	30	
-4	4	4	4	4	4	40	40	
-5	5	5	5	5	5	50	50	
-6	666	666	0	1234	\N	30	30	3,4,5,7
-
--- !seq_map_col_no_default_val_multi_rows_3 --
-0	0	0	0	0	0	0	
-1	888	888	1	1	1	20	3,4,5,7
-2	2	2	777	777	777	20	1,2,6,7,11
-3	777	777	555	555	555	53	7
-4	123	123	123	987	123	40	6,7,11
-5	5	5	5	5	5	50	
-6	666	666	0	1234	\N	30	3,4,5,7
-7	666	666	0	1234	\N	70	3,4,5,7
-
--- !seq_map_col_has_default_val_multi_rows_1 --
-0	0	0	0	0	0	0	0	
-1	1	1	1	1	1	10	10	
-2	2	2	2	2	2	20	20	
-3	3	3	3	3	3	30	30	
-4	4	4	4	4	4	40	40	
-5	5	5	5	5	5	50	50	
-
--- !seq_map_col_has_default_val_multi_rows_2 --
-0	0	0	0	0	0	0	0	
-1	888	888	1	1	1	20	20	3,4,5,7
-2	2	2	777	777	777	20	20	1,2,6,7,11
-3	3	3	3	3	3	30	30	
-4	4	4	4	4	4	40	40	
-5	5	5	5	5	5	50	50	
-6	123	123	0	666	666	60	60	3,6,7,11
-
--- !seq_map_col_has_default_val_multi_rows_3 --
-0	0	0	0	0	0	0	
-1	888	888	1	1	1	20	3,4,5,7
-2	2	2	777	777	777	20	1,2,6,7,11
-3	777	777	555	555	555	53	7
-4	123	123	123	987	123	40	6,7,11
-5	5	5	5	5	5	50	
-6	123	123	0	666	666	60	3,6,7,11
-7	666	666	0	1234	\N	70	3,4,5,7
-
